requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
dsnparse==0.2.1
exceptiongroup==1.2.1
Flask==3.0.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
sniffio==1.3.1
SQLAlchemy==2.0.27
typing_extensions==4.10.0
Werkzeug==3.0.1
